% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggical-package.R
\docType{package}
\name{ggical-package}
\alias{ggical}
\alias{ggical-package}
\title{ggical: Plot Day Plans from iCalendar Files}
\description{
Provides a 'ggplot2' geom, stat and theme to plot the time of day of events. Designed to work with 'ical' dataframes generated from the 'calendar' package. Provides reverse date, datetime, and time 'ggplot2' scales.
}
\author{
\strong{Maintainer}: Harry Thompson \email{harry@mayesfield.uk}

}
\keyword{internal}
