% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_planner.R
\name{theme_planner}
\alias{theme_planner}
\title{Planner Theme}
\usage{
theme_planner(
  orientation = c("x", "y"),
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{orientation}{which major grid lines should be removed?}

\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
A modification of the \code{\link[ggplot2:ggtheme]{ggplot2::theme_linedraw()}} complete theme.
Intended for use with \code{geom_planner()}.
}
\examples{
library(ggplot2)

calendar <- timetable
calendar$date <- as.Date(calendar$DTSTART)
today <- as.Date("2024-11-24")                # Sys.Date()
now <- as.POSIXct("2024-11-24 20:20:00 GMT")  # Sys.time()

ggplot(calendar) +
  geom_planner(aes(date, ystart = DTSTART, ystop = DTEND, fill = SUMMARY)) +
  geom_planner(aes(today, ystart = now), col = "red", linewidth = 0.1) +
  scale_x_date(
    date_breaks = "1 day", labels = ~weekdays(.x, abbreviate = TRUE),
    limits = today + c(-0.5, 6.5)) +
  scale_y_time_reverse(
    breaks = hms::hms(hours = 21:9), labels = ~format(as.POSIXlt(.x), format = "\%H:\%M"),
    limits = hms::hms(hours = c(21, 9))) +
  theme_planner()

ggplot(calendar) +
  geom_planner(aes(y = date, xstart = DTSTART, xstop = DTEND, fill = SUMMARY)) +
  geom_planner(aes(y = today, xstart = now), col = "red", linewidth = 0.1) +
  scale_x_time(
    breaks = hms::hms(hours = 9:21), labels = ~format(as.POSIXlt(.x), format = "\%H:\%M"),
    limits = hms::hms(hours = c(9, 21)), guide = guide_axis(angle = 45)) +
  scale_y_date_reverse(
    date_breaks = "-1 day", labels = weekdays,
    limits = today + c(6.5, -0.5)) +
  theme_planner("y")

}
