% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetable.R
\docType{data}
\name{timetable}
\alias{timetable}
\title{Student Timetable}
\format{
An object of class \code{ical} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 9 rows and 5 columns.
}
\source{
Fabricated.
}
\usage{
timetable
}
\description{
An example output of \code{\link[calendar:ic_read]{calendar::ic_read()}}.
}
\keyword{datasets}
